requirements.txt
openai==0.27.4
aiogram==2.25.1
python-dotenv==1.0.0